Support Iceberg partition compensation for mv rewrite
